+++ src/accessories/smarthome/smarthome-button.handler.js
@@ -108,12 +108,12 @@ class Handler {
                   let lowBattery = device.battery.low || 0;
 
                   accessory
-                    .getService(this.api.hap.Service.BatteryService)
+                    .getService(this.api.hap.Service.Battery)
                     .getCharacteristic(this.api.hap.Characteristic.BatteryLevel)
                     .updateValue(batteryLevel);
 
                   accessory
-                    .getService(this.api.hap.Service.BatteryService)
+                    .getService(this.api.hap.Service.Battery)
                     .getCharacteristic(this.api.hap.Characteristic.StatusLowBattery)
                     .updateValue(lowBattery);
                 } else {
@@ -250,12 +250,12 @@ class Handler {
                   let lowBattery = device.battery.low || 0;
 
                   accessory
-                    .getService(this.api.hap.Service.BatteryService)
+                    .getService(this.api.hap.Service.Battery)
                     .getCharacteristic(this.api.hap.Characteristic.BatteryLevel)
                     .updateValue(batteryLevel);
 
                   accessory
-                    .getService(this.api.hap.Service.BatteryService)
+                    .getService(this.api.hap.Service.Battery)
                     .getCharacteristic(this.api.hap.Characteristic.StatusLowBattery)
                     .updateValue(lowBattery);
                 } else {